Load Balancing System using Java
This project is implementation of the Load balancing system using Java
AWT. This speaks about the how to balance the load and handle the
requests are the coming from different clients.
This is in good condition.
To know the execution steps go to the documentation word file in the
project